refactor: Double → Long(amountInCents) 全链路改造
- Models.kt: Transaction.amount/MonthlySummary/YearlySummary 改为 Long（单位：分）
- TransactionDao.kt: 所有 SUM 返回 Long，search 参数 minAmount/maxAmount 改 Long
- YearDatabase.kt: version 1→2，MIGRATION_1_2 临时表迁移（旧 Double×100→Long）
- TransactionRepository.kt: 所有金额方法返回 Long
- SettingsRepository.kt: savingsBalance 改 Long
- MainViewModel.kt: StateFlow<Long>，getYearlySummary/getYearlySavings 返回 Long
- CommonComponents.kt: SummaryCard/SummaryItem 接收 Long，显示时 /100.0
- AddEditTransactionScreen.kt: 输入元→存储分，编辑回填分→元
- StatisticsScreen.kt: mutableLongStateOf，YearStatItem 接收 Long
- SearchScreen.kt: 金额搜索输入元→转分传参
- DatabaseManager.kt: performMigration 旧库 Double×100→Long

+++ app/src/main/java/com/simplebookkeeper/data/dao/TransactionDao.kt
@@ -58,59 +58,59 @@ interface TransactionDao {
     @Query("SELECT * FROM transactions ORDER BY date DESC LIMIT :limit")
     fun getRecent(limit: Int = 20): Flow<List<Transaction>>
 
-    // 本月收入总额
+    // 本月收入总额（单位：分）
     @Query("""
-        SELECT COALESCE(SUM(amount), 0) FROM transactions 
+        SELECT COALESCE(SUM(amount), 0) FROM transactions
         WHERE type = 'INCOME'
           AND strftime('%Y', date / 1000, 'unixepoch', 'localtime') = :year
           AND strftime('%m', date / 1000, 'unixepoch', 'localtime') = :month
     """)
-    fun getMonthlyIncome(year: String, month: String): Flow<Double>
+    fun getMonthlyIncome(year: String, month: String): Flow<Long>
 
-    // 本月支出总额
+    // 本月支出总额（单位：分）
     @Query("""
-        SELECT COALESCE(SUM(amount), 0) FROM transactions 
+        SELECT COALESCE(SUM(amount), 0) FROM transactions
         WHERE type = 'EXPENSE'
           AND strftime('%Y', date / 1000, 'unixepoch', 'localtime') = :year
           AND strftime('%m', date / 1000, 'unixepoch', 'localtime') = :month
     """)
-    fun getMonthlyExpense(year: String, month: String): Flow<Double>
+    fun getMonthlyExpense(year: String, month: String): Flow<Long>
 
-    // 某年收入总额
+    // 某年收入总额（单位：分）
     @Query("""
-        SELECT COALESCE(SUM(amount), 0) FROM transactions 
+        SELECT COALESCE(SUM(amount), 0) FROM transactions
         WHERE type = 'INCOME'
           AND strftime('%Y', date / 1000, 'unixepoch', 'localtime') = :year
     """)
-    suspend fun getYearlyIncome(year: String): Double
+    suspend fun getYearlyIncome(year: String): Long
 
-    // 某年支出总额
+    // 某年支出总额（单位：分）
     @Query("""
-        SELECT COALESCE(SUM(amount), 0) FROM transactions 
+        SELECT COALESCE(SUM(amount), 0) FROM transactions
         WHERE type = 'EXPENSE'
           AND strftime('%Y', date / 1000, 'unixepoch', 'localtime') = :year
     """)
-    suspend fun getYearlyExpense(year: String): Double
+    suspend fun getYearlyExpense(year: String): Long
 
-    // 某年储蓄总额
+    // 某年储蓄总额（单位：分）
     @Query("""
-        SELECT COALESCE(SUM(amount), 0) FROM transactions 
+        SELECT COALESCE(SUM(amount), 0) FROM transactions
         WHERE type = 'SAVING'
           AND strftime('%Y', date / 1000, 'unixepoch', 'localtime') = :year
     """)
-    suspend fun getYearlySavingAmount(year: String): Double
+    suspend fun getYearlySavingAmount(year: String): Long
 
-    // 某年支取总额
+    // 某年支取总额（单位：分）
     @Query("""
-        SELECT COALESCE(SUM(amount), 0) FROM transactions 
+        SELECT COALESCE(SUM(amount), 0) FROM transactions
         WHERE type = 'WITHDRAW'
           AND strftime('%Y', date / 1000, 'unixepoch', 'localtime') = :year
     """)
-    suspend fun getYearlyWithdrawAmount(year: String): Double
+    suspend fun getYearlyWithdrawAmount(year: String): Long
 
-    // 模糊查询（时间范围 + 金额 + 类型 + 分类 + 备注关键词）
+    // 模糊查询（时间范围 + 金额(分) + 类型 + 分类 + 备注关键词）
     @Query("""
-        SELECT * FROM transactions 
+        SELECT * FROM transactions
         WHERE (:startDate IS NULL OR date >= :startDate)
           AND (:endDate IS NULL OR date <= :endDate)
           AND (:minAmount IS NULL OR amount >= :minAmount)
@@ -123,8 +123,8 @@ interface TransactionDao {
     fun search(
         startDate: Long? = null,
         endDate: Long? = null,
-        minAmount: Double? = null,
-        maxAmount: Double? = null,
+        minAmount: Long? = null,
+        maxAmount: Long? = null,
         type: String? = null,
         categoryId: Long? = null,
         keyword: String? = null
